Created a Rock Paper Scissors Game
Checked to see if the user or computer won or if it was a tie. It kept going in a while loop. The only time it ends is if the user types q. In the end it will tell you how many points the user and the computer got. The values have been stored in a list.